# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/fulfillment": {
      "get": {
        "operationId": "AdminFulfillmentController_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Get admin fulfillment settings and readiness"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-fulfillment"
        ]
      }
    },
    "/admin/fulfillment/quick-setup": {
      "post": {
        "operationId": "AdminFulfillmentController_quickSetup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickFulfillmentSetupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create suggested fulfillment setup"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-fulfillment"
        ]
      }
    },
    "/admin/fulfillment/zones": {
      "post": {
        "operationId": "AdminFulfillmentController_createZone",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShippingZoneDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create fulfillment zone"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-fulfillment"
        ]
      }
    },
    "/admin/fulfillment/zones/{zoneId}": {
      "put": {
        "operationId": "AdminFulfillmentController_updateZone",
        "parameters": [
          {
            "name": "zoneId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShippingZoneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update fulfillment zone"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-fulfillment"
        ]
      },
      "delete": {
        "operationId": "AdminFulfillmentController_removeZone",
        "parameters": [
          {
            "name": "zoneId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Delete fulfillment zone"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-fulfillment"
        ]
      }
    },
    "/admin/fulfillment/zones/{zoneId}/methods": {
      "post": {
        "operationId": "AdminFulfillmentController_createMethod",
        "parameters": [
          {
            "name": "zoneId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShippingMethodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create fulfillment method"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-fulfillment"
        ]
      }
    },
    "/admin/fulfillment/zones/{zoneId}/methods/{methodId}": {
      "put": {
        "operationId": "AdminFulfillmentController_updateMethod",
        "parameters": [
          {
            "name": "zoneId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "methodId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShippingMethodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update fulfillment method"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-fulfillment"
        ]
      },
      "delete": {
        "operationId": "AdminFulfillmentController_removeMethod",
        "parameters": [
          {
            "name": "zoneId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "methodId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Delete fulfillment method"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-fulfillment"
        ]
      }
    },
...
  },
  "info": {
    "title": "Alnjoom Telecom Store API",
    "description": "Backend APIs for Alnjoom Telecom Store",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateShippingZoneDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateShippingZoneDto": {
        "type": "object",
        "properties": {}
      },
      "CreateShippingMethodDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateShippingMethodDto": {
        "type": "object",
        "properties": {}
      },
      "QuickFulfillmentSetupDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}